Fix DateTime64 pre-epoch support in Sparkbar combinator
Pre-epoch DateTime64 timestamps have negative raw tick counts. The previous
code used UInt64 as Key for DateTime64, so negative begin_x/end_x were cast
to large unsigned values, causing the begin_x >= end_x constructor check to
throw BAD_ARGUMENTS for valid cross-epoch ranges.

Fix: use Int64 as Key for DateTime64 (same as Int*), keeping tick counts
signed end-to-end. The UInt64 subtraction trick in add() is unaffected:
after the key >= begin_x bounds check, both deltas are mathematically
non-negative, so two's complement wraparound gives correct results.

Add a cross-epoch test (begin_x=-2000ms, end_x=+2000ms) to prevent
regression.

+++ src/AggregateFunctions/Combinators/AggregateFunctionSparkbar.cpp
@@ -93,12 +93,13 @@ public:
 
         if (which.isDateTime64())
         {
-            /// DateTime64 parameters are DecimalField<Decimal64>; extract .value directly to preserve raw ticks.
-            const auto extract = [](const Field & f) -> UInt64
+            /// DateTime64 parameters are DecimalField<DateTime64>; extract raw ticks as Int64
+            /// to correctly handle pre-epoch timestamps (negative tick counts).
+            const auto extract = [](const Field & f) -> Int64
             {
-                return static_cast<UInt64>(static_cast<Int64>(f.safeGet<DecimalField<DateTime64>>().getValue()));
+                return static_cast<Int64>(f.safeGet<DecimalField<DateTime64>>().getValue());
             };
-            return std::make_shared<AggregateFunctionSparkbar<UInt64>>(
+            return std::make_shared<AggregateFunctionSparkbar<Int64>>(
                 nested_function, width, extract(params[1]), extract(params[2]), arguments, params);
         }
